// src/types/planets.ts
export interface planetInfo {
    distance_light_year: number
    host_star_mass: number
    host_star_temperature: number
    mass: number
    name: string
    period: number
    radius: number
    semi_major_axis: number
    temperature: number
}

export type IPlanetNames  =  "Mercury" | "Venus" | "Earth" | "Mars" | "Jupiter" | "Saturn" | "Uranus" | "Neptune"